# openapi.yaml
components:
  schemas:
    APIException:
      additionalProperties: true
      properties:
        detail:
          title: Detail
          type: string
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
        status:
          title: Status
          type: integer
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      required:
      - status
      - detail
      title: APIException
      type: object
    AdditionalParameter:
      properties:
        name:
          title: Name
          type: string
        value:
          items:
            anyOf:
            - type: string
            - type: number
            - type: integer
            - items:
                type: object
              type: array
            - type: object
          title: Value
          type: array
      required:
      - name
      - value
      title: AdditionalParameter
      type: object
    AdditionalParameters:
      properties:
        href:
          anyOf:
          - type: string
          - type: 'null'
          title: Href
        parameters:
          anyOf:
          - items:
              $ref: '#/components/schemas/AdditionalParameter'
            type: array
          - type: 'null'
          title: Parameters
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      required:
      - title
      - role
      - href
      - parameters
      title: AdditionalParameters
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    InputDescription:
      properties:
        additionalParameters:
          anyOf:
          - $ref: '#/components/schemas/AdditionalParameters'
          - type: 'null'
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
        maxOccurs:
          anyOf:
          - minimum: 0.0
            type: integer
          - $ref: '#/components/schemas/MaxOccur'
          default: unbounded
          title: Maxoccurs
        metadata:
          anyOf:
          - items:
              $ref: '#/components/schemas/Metadata'
            type: array
          - type: 'null'
          title: Metadata
        minOccurs:
          default: 1
          minimum: 0.0
          title: Minoccurs
          type: integer
        schema:
          $ref: '#/components/schemas/Reference'
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      required:
      - title
      - description
      - keywords
      - metadata
      - additionalParameters
      - minOccurs
      - maxOccurs
      - schema
      title: InputDescription
      type: object
...
    JobControlOptions:
      const: async-execute
      title: JobControlOptions
      type: string
...
    Link-Input:
      properties:
        href:
          format: uri
          minLength: 1
          title: Href
          type: string
        hreflang:
          anyOf:
          - type: string
          - type: 'null'
          title: Hreflang
        rel:
          title: Rel
          type: string
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        type:
          title: Type
          type: string
      required:
      - href
      - rel
      - type
      title: Link
      type: object
    Link-Output:
      properties:
        href:
          format: uri
          minLength: 1
          title: Href
          type: string
        hreflang:
          anyOf:
          - type: string
          - type: 'null'
          title: Hreflang
        rel:
          title: Rel
          type: string
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        type:
          title: Type
          type: string
      required:
      - href
      - rel
      - type
      - hreflang
      - title
      title: Link
      type: object
    MaxOccur:
      const: unbounded
      title: MaxOccur
      type: string
    Metadata:
      properties:
        href:
          anyOf:
          - type: string
          - type: 'null'
          title: Href
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      required:
      - title
      - role
      - href
      title: Metadata
      type: object
    OutputDescription:
      properties:
        additionalParameters:
          anyOf:
          - $ref: '#/components/schemas/AdditionalParameters'
          - type: 'null'
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
        metadata:
          anyOf:
          - items:
              $ref: '#/components/schemas/Metadata'
            type: array
          - type: 'null'
          title: Metadata
        schema:
          $ref: '#/components/schemas/Reference'
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      required:
      - title
      - description
      - keywords
      - metadata
      - additionalParameters
      - schema
      title: OutputDescription
      type: object
...
    Process:
      properties:
        additionalParameters:
          anyOf:
          - $ref: '#/components/schemas/AdditionalParameters'
          - type: 'null'
        cacheKeyHashExcludes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Cachekeyhashexcludes
        cacheKeyHashIncludes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Cachekeyhashincludes
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        handlerType:
          title: Handlertype
          type: string
        id:
          title: Id
          type: string
        inputs:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/InputDescription'
            type: object
          - type: 'null'
          title: Inputs
        jobControlOptions:
          anyOf:
          - items:
              $ref: '#/components/schemas/JobControlOptions'
            type: array
          - type: 'null'
          default:
          - async-execute
          title: Jobcontroloptions
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
        links:
          default: []
          items:
            $ref: '#/components/schemas/Link-Output'
          title: Links
          type: array
        metadata:
          anyOf:
          - items:
              $ref: '#/components/schemas/Metadata'
            type: array
          - type: 'null'
          title: Metadata
        outputTransmission:
          anyOf:
          - items:
              $ref: '#/components/schemas/TransmissionMode'
            type: array
          - type: 'null'
          title: Outputtransmission
        outputs:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/OutputDescription'
            type: object
          - type: 'null'
          title: Outputs
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        version:
          title: Version
          type: string
      required:
      - title
      - description
      - keywords
      - metadata
      - additionalParameters
      - id
      - version
      - handlerType
      - jobControlOptions
      - outputTransmission
      - links
      - inputs
      - outputs
      - cacheKeyHashIncludes
      - cacheKeyHashExcludes
      title: Process
      type: object
    ProcessList:
      properties:
        links:
          items:
            $ref: '#/components/schemas/Link-Output'
          title: Links
          type: array
        processes:
          items:
            $ref: '#/components/schemas/ProcessSummary'
          title: Processes
          type: array
      required:
      - processes
      - links
      title: ProcessList
      type: object
    ProcessSummary:
      properties:
        additionalParameters:
          anyOf:
          - $ref: '#/components/schemas/AdditionalParameters'
          - type: 'null'
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        handlerType:
          title: Handlertype
          type: string
        id:
          title: Id
          type: string
        jobControlOptions:
          anyOf:
          - items:
              $ref: '#/components/schemas/JobControlOptions'
            type: array
          - type: 'null'
          default:
          - async-execute
          title: Jobcontroloptions
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
        links:
          default: []
          items:
            $ref: '#/components/schemas/Link-Output'
          title: Links
          type: array
        metadata:
          anyOf:
          - items:
              $ref: '#/components/schemas/Metadata'
            type: array
          - type: 'null'
          title: Metadata
        outputTransmission:
          anyOf:
          - items:
              $ref: '#/components/schemas/TransmissionMode'
            type: array
          - type: 'null'
          title: Outputtransmission
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        version:
          title: Version
          type: string
      required:
      - title
      - description
      - keywords
      - metadata
      - additionalParameters
      - id
      - version
      - handlerType
      - jobControlOptions
      - outputTransmission
      - links
      title: ProcessSummary
      type: object
...
    Reference:
      additionalProperties: false
      properties:
        $ref:
          title: $Ref
          type: string
      required:
      - $ref
      title: Reference
      type: object
...
    StatusCode:
      enum:
      - accepted
      - running
      - successful
      - failed
      - dismissed
      title: StatusCode
      type: string
    StatusInfo-Input:
      properties:
        created:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created
        finished:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished
        jobID:
          format: uuid
          title: Jobid
          type: string
        links:
          default: []
          items:
            $ref: '#/components/schemas/Link-Input'
          title: Links
          type: array
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        processID:
          title: Processid
          type: string
        started:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Started
        status:
          $ref: '#/components/schemas/StatusCode'
        type:
          $ref: '#/components/schemas/Type'
        updated:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated
      required:
      - processID
      - type
      - jobID
      - status
      title: StatusInfo
      type: object
...
    TransmissionMode:
      enum:
      - value
      - reference
      title: TransmissionMode
      type: string
    Type:
      const: process
      title: Type
      type: string
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: swoop-api
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /processes:
    get:
      description: Returns a list of all available workflows
      operationId: list_workflows_processes_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 1000
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: handler
        required: false
        schema:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Handler
      - in: query
        name: type
        required: false
        schema:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Type
      - in: query
        name: lastID
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessList'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Workflows
      tags:
      - Processes
  /processes/{processID}:
    get:
      description: Returns workflow details by processID
      operationId: get_workflow_description_processes__processID__get
      parameters:
      - in: path
        name: processID
        required: true
        schema:
          title: Processid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Internal Server Error
      summary: Get Workflow Description
      tags:
      - Processes
  /processes/{processID}/execution:
    post:
      description: Starts a workflow execution (Job)
      operationId: execute_workflow_processes__processID__execution_post
      parameters:
      - in: path
        name: processID
        required: true
        schema:
          title: Processid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              title: Body
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusInfo-Input'
          description: Successful Response
        '303':
          description: See existing job
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Unprocessable Entity
      summary: Execute Workflow
      tags:
      - Processes
  /processes/{processID}/inputsschema:
    get:
      description: Returns process input jsonschema by processID
      operationId: get_process_inputs_schema_processes__processID__inputsschema_get
      parameters:
      - in: path
        name: processID
        required: true
        schema:
          title: Processid
      responses:
        '200':
          content:
            application/json:
              schema: {}
            application/schema+json: {}
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Internal Server Error
      summary: Get Process Inputs Schema
      tags:
      - Processes
  /processes/{processID}/outputsschema:
    get:
      description: Returns workflow input jsonschema by processID
      operationId: get_process_outputs_schema_processes__processID__outputsschema_get
      parameters:
      - in: path
        name: processID
        required: true
        schema:
          title: Processid
      responses:
        '200':
          content:
            application/json:
              schema: {}
            application/schema+json: {}
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIException'
          description: Internal Server Error
      summary: Get Process Outputs Schema
      tags:
      - Processes